move item-related functions from proc to feed module
